Added logging, option to delete command, tag users on expansion info errors, started error checking, moved table classes into their own package, new gary_messages table for storing messages, fixed requestutils not picking up requests when markdown was used.

+++ src/main/java/me/piggypiglet/gary/core/handlers/CommandHandler.java
@@ -46,7 +46,10 @@ public final class CommandHandler extends ListenerAdapter {
 
                 if (mutil.startsWith(msg, name)) {
                     command.run(e, args);
-                    e.getMessage().delete().queue();
+
+                    if (command.getDelete()) {
+                        e.getMessage().delete().queue();
+                    }
                 }
             }
             checkMessage(e);
